requirements.txt
flask
numpy
pillow
onnxruntime
gunicorn